+++ projects/worldskills-angular-lib/src/lib/cache/models/cache-wrapper.ts
@@ -1,0 +1,5 @@
+export interface CacheWrapper {
+    data: any;
+    expire?: Date;
+  }
+